work on placing items
