+++ Assets/Prefabs/Catapult.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9ab5556371c30a64aa8a406c1e000fc7
+guid: 8d65069d52fb5c247b1caccb9d3003ff
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
