+++ Assets/Prefabs/Catapult/Catapult Variant.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 809f0fade5c383849abc60c469350ba0
-folderAsset: yes
-DefaultImporter:
+guid: 116445ef4d88d224ba0204e76c5dfd1f
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
