+++ Assets/Scenes/DevAssets/Catapult/Catapult-Inventory.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4d1d6ff47bad92141be63115e4144121
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
